chebul pings purged

+++ cogs/Cherry Bullet.py
@@ -11,9 +11,6 @@ sadboi = 642497143801905190
 #.luminary bot-commands
 kbotcom = 764610881513324574
 
-#//people
-ple = 416903886968979466
-
 class CherryBullet(commands.Cog):
 
     def __init__(self, bot):
@@ -264,14 +261,9 @@ class CherryBullet(commands.Cog):
                 await ctx.send(random.choice(self.bot.cherrybullet_haeyoon_gif))
                 await ctx.message.delete()
         elif arg == "jiwon" or arg == "bullet jiwon":
-            if ctx.guild.id == luminary:
-                if ctx.channel.id == kbotcom:
-                    await ctx.send(f'<@{ple}>, <@!{ctx.author.id}> is talking about Jiwon :heart:')
-                    await ctx.send(random.choice(self.bot.cherrybullet_jiwon_gif))
-                    await ctx.message.delete()
-                else:
-                    await ctx.send(content=f'Wrong channel <@!{ctx.author.id}>! Go to <#764610881513324574>', delete_after=2)
-                    await ctx.message.delete()
+            if ctx.guild.id == luminary and ctx.channel.id != kbotcom:
+                await ctx.send(content=f'Wrong channel <@!{ctx.author.id}>! Go to <#764610881513324574>', delete_after=2)
+                await ctx.message.delete()
             else:
                 await ctx.send(f'<@!{ctx.author.id}> is talking about Jiwon :heart:')
                 await ctx.send(random.choice(self.bot.cherrybullet_jiwon_gif))
